refactor - make new app for talks
