+++ chipy_org/apps/meetings/urls.py
@@ -6,11 +6,7 @@ from .views import (
     RSVP,
     FutureMeetings,
     MeetingDetail,
-    MyTopics,
     PastMeetings,
-    PastTopic,
-    PastTopics,
-    ProposeTopic,
     RSVPlistHost,
     RSVPlistPrivate,
     UpcomingEvents,
@@ -35,9 +31,5 @@ urlpatterns = [
         RSVPlistHost.as_view(),
         name="rsvp_list_host_csv",
     ),
-    path("topics/propose/", login_required(ProposeTopic.as_view()), name="propose_topic"),
-    path("topics/mine/", login_required(MyTopics.as_view()), name="my_topics"),
-    path("topics/past/", PastTopics.as_view(), name="past_topics"),
-    path("topics/past/<int:id>/", PastTopic.as_view(), name="past_topic"),
     path("upcoming_events/", UpcomingEvents.as_view(), name="upcoming_events"),
 ]
