+++ chipy_org/apps/talks/templates/talks/_talk.html
