+++ chipy_org/apps/talks/templates/talks/past_topics.html
@@ -18,7 +18,7 @@
 <div class="row">
   <div>
     {% for talk in topics %}
-      {% include 'meetings/_talk.html' %}
+      {% include 'talks/_talk.html' %}
     {% endfor %}
     <span class="count">{{ object_list|length }} talk{{ object_list|pluralize }}</span>
   </div>
